Delete SimpleSchema/SimpleObj and Move schemas to dedicated files
