+++ packages/backend/src/models/schema/note-reaction.ts
@@ -1,0 +1,26 @@
+export const packedNoteReactionSchema = {
+	type: 'object' as const,
+	optional: false as const, nullable: false as const,
+	properties: {
+		id: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+			format: 'id',
+			example: 'xxxxxxxxxx',
+		},
+		createdAt: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+			format: 'date-time',
+		},
+		user: {
+			type: 'object' as const,
+			optional: false as const, nullable: false as const,
+			ref: 'User' as const,
+		},
+		type: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+		},
+	},
+};
