+++ packages/backend/src/models/schema/page.ts
@@ -1,0 +1,52 @@
+export const packedPageSchema = {
+	type: 'object' as const,
+	optional: false as const, nullable: false as const,
+	properties: {
+		id: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+			format: 'id',
+			example: 'xxxxxxxxxx',
+		},
+		createdAt: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+			format: 'date-time',
+		},
+		updatedAt: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+			format: 'date-time',
+		},
+		title: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+		},
+		name: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+		},
+		summary: {
+			type: 'string' as const,
+			optional: false as const, nullable: true as const,
+		},
+		content: {
+			type: 'array' as const,
+			optional: false as const, nullable: false as const,
+		},
+		variables: {
+			type: 'array' as const,
+			optional: false as const, nullable: false as const,
+		},
+		userId: {
+			type: 'string' as const,
+			optional: false as const, nullable: false as const,
+			format: 'id',
+		},
+		user: {
+			type: 'object' as const,
+			ref: 'User' as const,
+			optional: false as const, nullable: false as const,
+		},
+	},
+};
